moved query parser to http/query/internal instead of internal/query
