+++ http/query/internal/parserbench_test.go
@@ -1,4 +1,4 @@
-package query
+package internal
 
 import (
 	"fmt"
